+++ Frontend/src/pages/GestionPedidosPage.tsx
@@ -571,12 +571,13 @@ const GestionPedidosPage: React.FC = () => {
   /**
    * @brief Calcula el total de ventas del día, lo guarda en el estado y abre el modal.
    */
-  const handleCalcularTotalVentas = useCallback(() => {
-    const totalDelDia = getIngresosBrutosByDate(searchDate);
-
-    setTotalVentasDia(totalDelDia.total);
+  const handleCalcularTotalVentas = useCallback(async () => {
+    const data = await getIngresosBrutosByDate(searchDate);
+    console.log("Datos recibidos:", data);
+    const totalDelDia: number = Number(data.total)
+    setTotalVentasDia(totalDelDia);
     setIsTotalVentasModalOpen(true);
-  }, [pedidos]); // Se recalcula si la lista de pedidos cambia
+  }, [searchDate]); // Se recalcula si la lista de pedidos cambia
 
   /**
    * @brief Cierra el modal de total de ventas.
@@ -853,7 +854,7 @@ const GestionPedidosPage: React.FC = () => {
             <div className={styles.detailSection}>
               <p><strong>Fecha:</strong> {new Date(searchDate + 'T00:00:00').toLocaleDateString('es-AR', { day: '2-digit', month: '2-digit', year: 'numeric' })}</p>
               <p style={{ fontSize: '1.8rem', marginTop: '20px' }}>
-                <strong>Total: ${totalVentasDia.toFixed(2)}</strong>
+                <strong>Total: ${Number(totalVentasDia).toFixed(0)}</strong>
               </p>
             </div>
             <div className={styles.modalActions}>